Safari-611.1.10/LayoutTests/editing/deleting/delete-line-break-before-underlined-content.html

<!DOCTYPE html>
<html>
<body>
<div id="container" contenteditable="true">This shouldn't be underlined.<div id="div" style="text-decoration: underline;">This should be underlined.</div></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
var div = document.getElementById("div");
var range = document.createRange();
range.setStart(div, 0);
range.setEnd(div, 0);
var sel = window.getSelection();
sel.addRange(range);
document.execCommand("Delete", false, null);

Markup.description('This tests for a bug where underlined content would lose its underlining' +
    'when deleting the line break before the paragraph that contained it.');

Markup.dump('container');
</script>
</body>
</html>
