LayoutTests/editing/deleting/delete-line-break-before-underlined-content-expected.txt

This tests for a bug where underlined content would lose its underliningwhen deleting the line break before the paragraph that contained it.
| "This shouldn't be underlined.<#selection-caret>"
| <div>
|   id="div"
|   style="text-decoration: underline; display: inline !important; "
|   "This should be underlined."
